Update formatting to fit pass-off program

+++ CommentAutomaton.cpp
@@ -50,7 +50,7 @@ void CommentAutomaton::S2(const std::string &input) {
 }
 
 void CommentAutomaton::SB1(const std::string &input) {
-    if (input.size() <= index) { // Reached end of string
+    if ((int)input.size() <= index) { // Reached end of string
         Serr();
     }
     else if (input[index] == '|') {
@@ -72,7 +72,7 @@ void CommentAutomaton::SB1(const std::string &input) {
 }
 
 void CommentAutomaton::SB2(const std::string &input){
-    if (input.size() <= index) { // end of file
+    if ((int)input.size() <= index) { // end of file
         Serr();
     }
     else if (input[index] == '#') { // end of comment
